+++ NHRMS_WebAPI/Models/EmployeeBranch.cs
@@ -9,7 +9,7 @@ namespace NHRMS_WebAPI.Models
     {
         public long ID { get; set; }
         public long EmployeeID { get; set; }
-        public int BranchID { get; set; }
+        public long BranchID { get; set; }
         public string BranchName { get; set; }
         public bool IsActive { get; set; }
     }
